requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==4.2.20
django-allauth==65.4.1
django-environ==0.12.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
gunicorn==20.1.0
whitenoise==6.4.0
django-paypal
celery==5.4.0
redis==4.6.0
